app/requirements.txt
fastapi
uvicorn
pillow
onnxruntime
